style(all): Rearrange the file directory and update makefile.
