+++ .gitignore
@@ -20,6 +20,7 @@ target/*
 *.apk
 cocktail
 cocktail.exe
+*.o
 
 # Project files, i.e. `.project`, `.actionScriptProperties` and `.flexProperties`
 # should NOT be excluded as they contain compiler settings and other important
