--- Source/cmake-build-debug/CMakeFiles/3.13.4/CMakeSystem.cmake
@@ -1,15 +1,0 @@
-set(CMAKE_HOST_SYSTEM "Linux-4.4.0-19041-Microsoft")
-set(CMAKE_HOST_SYSTEM_NAME "Linux")
-set(CMAKE_HOST_SYSTEM_VERSION "4.4.0-19041-Microsoft")
-set(CMAKE_HOST_SYSTEM_PROCESSOR "x86_64")
-
-
-
-set(CMAKE_SYSTEM "Linux-4.4.0-19041-Microsoft")
-set(CMAKE_SYSTEM_NAME "Linux")
-set(CMAKE_SYSTEM_VERSION "4.4.0-19041-Microsoft")
-set(CMAKE_SYSTEM_PROCESSOR "x86_64")
-
-set(CMAKE_CROSSCOMPILING "FALSE")
-
-set(CMAKE_SYSTEM_LOADED 1)
