--- Source/cmake-build-debug/CMakeFiles/Makefile.cmake
@@ -1,46 +1,0 @@
-# CMAKE generated file: DO NOT EDIT!
-# Generated by "Unix Makefiles" Generator, CMake Version 3.13
-
-# The generator used is:
-set(CMAKE_DEPENDS_GENERATOR "Unix Makefiles")
-
-# The top level Makefile was generated from the following files:
-set(CMAKE_MAKEFILE_DEPENDS
-  "CMakeCache.txt"
-  "../CMakeLists.txt"
-  "CMakeFiles/3.13.4/CMakeCCompiler.cmake"
-  "CMakeFiles/3.13.4/CMakeSystem.cmake"
-  "/usr/share/cmake-3.13/Modules/CMakeCInformation.cmake"
-  "/usr/share/cmake-3.13/Modules/CMakeCommonLanguageInclude.cmake"
-  "/usr/share/cmake-3.13/Modules/CMakeExtraGeneratorDetermineCompilerMacrosAndIncludeDirs.cmake"
-  "/usr/share/cmake-3.13/Modules/CMakeFindCodeBlocks.cmake"
-  "/usr/share/cmake-3.13/Modules/CMakeGenericSystem.cmake"
-  "/usr/share/cmake-3.13/Modules/CMakeInitializeConfigs.cmake"
-  "/usr/share/cmake-3.13/Modules/CMakeLanguageInformation.cmake"
-  "/usr/share/cmake-3.13/Modules/CMakeSystemSpecificInformation.cmake"
-  "/usr/share/cmake-3.13/Modules/CMakeSystemSpecificInitialize.cmake"
-  "/usr/share/cmake-3.13/Modules/Compiler/CMakeCommonCompilerMacros.cmake"
-  "/usr/share/cmake-3.13/Modules/Compiler/GNU-C.cmake"
-  "/usr/share/cmake-3.13/Modules/Compiler/GNU.cmake"
-  "/usr/share/cmake-3.13/Modules/Platform/Linux-GNU-C.cmake"
-  "/usr/share/cmake-3.13/Modules/Platform/Linux-GNU.cmake"
-  "/usr/share/cmake-3.13/Modules/Platform/Linux.cmake"
-  "/usr/share/cmake-3.13/Modules/Platform/UnixPaths.cmake"
-  "/usr/share/cmake-3.13/Modules/ProcessorCount.cmake"
-  )
-
-# The corresponding makefile is:
-set(CMAKE_MAKEFILE_OUTPUTS
-  "Makefile"
-  "CMakeFiles/cmake.check_cache"
-  )
-
-# Byproducts of CMake generate step:
-set(CMAKE_MAKEFILE_PRODUCTS
-  "CMakeFiles/CMakeDirectoryInformation.cmake"
-  )
-
-# Dependency information for all targets:
-set(CMAKE_DEPEND_INFO_FILES
-  "CMakeFiles/cocktail.dir/DependInfo.cmake"
-  )
