--- Source/cmake-build-debug/CMakeFiles/TargetDirectories.txt
@@ -1,3 +1,0 @@
-/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug/CMakeFiles/rebuild_cache.dir
-/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug/CMakeFiles/cocktail.dir
-/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug/CMakeFiles/edit_cache.dir
