--- Source/cmake-build-debug/CMakeFiles/clion-log.txt
@@ -1,4 +1,0 @@
-/usr/bin/cmake -DCMAKE_BUILD_TYPE=Debug -DCMAKE_MAKE_PROGRAM=/usr/bin/make -DCMAKE_C_COMPILER=/usr/bin/gcc -DCMAKE_CXX_COMPILER=/usr/bin/g++ -G "CodeBlocks - Unix Makefiles" /tmp/tmp.RV7yFDR3Zp/Source
--- Configuring done
--- Generating done
--- Build files have been written to: /tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug
