--- Source/cmake-build-debug/CMakeFiles/cmake.check_cache
@@ -1,1 +1,0 @@
-# This file is generated by cmake for dependency checking of the CMakeCache.txt file
