--- Source/cmake-build-debug/CMakeFiles/cocktail.dir/C.includecache
@@ -1,84 +1,0 @@
-#IncludeRegexLine: ^[ 	]*[#%][ 	]*(include|import)[ 	]*[<"]([^">]+)([">])
-
-#IncludeRegexScan: ^.*$
-
-#IncludeRegexComplain: ^$
-
-#IncludeRegexTransform: 
-
-/tmp/tmp.RV7yFDR3Zp/Source/cocktail.h
-cot_util.h
-/tmp/tmp.RV7yFDR3Zp/Source/cot_util.h
-
-/tmp/tmp.RV7yFDR3Zp/Source/cot_debug.c
-cot_debug.h
-/tmp/tmp.RV7yFDR3Zp/Source/cot_debug.h
-
-/tmp/tmp.RV7yFDR3Zp/Source/cot_debug.h
-stdio.h
--
-
-/tmp/tmp.RV7yFDR3Zp/Source/cot_lexer.c
-cot_lexer.h
-/tmp/tmp.RV7yFDR3Zp/Source/cot_lexer.h
-
-/tmp/tmp.RV7yFDR3Zp/Source/cot_lexer.h
-stdio.h
--
-stdlib.h
--
-string.h
--
-cot_util.h
-/tmp/tmp.RV7yFDR3Zp/Source/cot_util.h
-cot_debug.h
-/tmp/tmp.RV7yFDR3Zp/Source/cot_debug.h
-cocktail.h
-/tmp/tmp.RV7yFDR3Zp/Source/cocktail.h
-
-/tmp/tmp.RV7yFDR3Zp/Source/cot_parser.c
-cocktail.h
-/tmp/tmp.RV7yFDR3Zp/Source/cocktail.h
-cot_parser.h
-/tmp/tmp.RV7yFDR3Zp/Source/cot_parser.h
-
-/tmp/tmp.RV7yFDR3Zp/Source/cot_parser.h
-cocktail.h
-/tmp/tmp.RV7yFDR3Zp/Source/cocktail.h
-cot_util.h
-/tmp/tmp.RV7yFDR3Zp/Source/cot_util.h
-stdlib.h
--
-
-/tmp/tmp.RV7yFDR3Zp/Source/cot_stack.c
-cot_stack.h
-/tmp/tmp.RV7yFDR3Zp/Source/cot_stack.h
-
-/tmp/tmp.RV7yFDR3Zp/Source/cot_stack.h
-cot_util.h
-/tmp/tmp.RV7yFDR3Zp/Source/cot_util.h
-
-/tmp/tmp.RV7yFDR3Zp/Source/cot_type.h
-
-/tmp/tmp.RV7yFDR3Zp/Source/cot_util.c
-cot_util.h
-/tmp/tmp.RV7yFDR3Zp/Source/cot_util.h
-
-/tmp/tmp.RV7yFDR3Zp/Source/cot_util.h
-stdio.h
--
-string.h
--
-cot_type.h
-/tmp/tmp.RV7yFDR3Zp/Source/cot_type.h
-
-/tmp/tmp.RV7yFDR3Zp/Source/main.c
-string.h
--
-cot_debug.h
-/tmp/tmp.RV7yFDR3Zp/Source/cot_debug.h
-cot_lexer.h
-/tmp/tmp.RV7yFDR3Zp/Source/cot_lexer.h
-cot_parser.h
-/tmp/tmp.RV7yFDR3Zp/Source/cot_parser.h
-
