--- Source/cmake-build-debug/CMakeFiles/cocktail.dir/DependInfo.cmake
@@ -1,25 +1,0 @@
-# The set of languages for which implicit dependencies are needed:
-set(CMAKE_DEPENDS_LANGUAGES
-  "C"
-  )
-# The set of files for implicit dependencies of each language:
-set(CMAKE_DEPENDS_CHECK_C
-  "/tmp/tmp.RV7yFDR3Zp/Source/cot_debug.c" "/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug/CMakeFiles/cocktail.dir/cot_debug.c.o"
-  "/tmp/tmp.RV7yFDR3Zp/Source/cot_lexer.c" "/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug/CMakeFiles/cocktail.dir/cot_lexer.c.o"
-  "/tmp/tmp.RV7yFDR3Zp/Source/cot_parser.c" "/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug/CMakeFiles/cocktail.dir/cot_parser.c.o"
-  "/tmp/tmp.RV7yFDR3Zp/Source/cot_stack.c" "/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug/CMakeFiles/cocktail.dir/cot_stack.c.o"
-  "/tmp/tmp.RV7yFDR3Zp/Source/cot_util.c" "/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug/CMakeFiles/cocktail.dir/cot_util.c.o"
-  "/tmp/tmp.RV7yFDR3Zp/Source/main.c" "/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug/CMakeFiles/cocktail.dir/main.c.o"
-  )
-set(CMAKE_C_COMPILER_ID "GNU")
-
-# The include file search paths:
-set(CMAKE_C_TARGET_INCLUDE_PATH
-  )
-
-# Targets to which this target links.
-set(CMAKE_TARGET_LINKED_INFO_FILES
-  )
-
-# Fortran module output directory.
-set(CMAKE_Fortran_TARGET_MODULE_DIR "")
