--- Source/cmake-build-debug/CMakeFiles/cocktail.dir/cmake_clean.cmake
@@ -1,15 +1,0 @@
-file(REMOVE_RECURSE
-  "CMakeFiles/cocktail.dir/cot_debug.c.o"
-  "CMakeFiles/cocktail.dir/cot_stack.c.o"
-  "CMakeFiles/cocktail.dir/cot_lexer.c.o"
-  "CMakeFiles/cocktail.dir/cot_parser.c.o"
-  "CMakeFiles/cocktail.dir/cot_util.c.o"
-  "CMakeFiles/cocktail.dir/main.c.o"
-  "cocktail.pdb"
-  "cocktail"
-)
-
-# Per-language clean rules from dependency scanning.
-foreach(lang C)
-  include(CMakeFiles/cocktail.dir/cmake_clean_${lang}.cmake OPTIONAL)
-endforeach()
