--- Source/cmake-build-debug/CMakeFiles/cocktail.dir/depend.internal
@@ -1,36 +1,0 @@
-# CMAKE generated file: DO NOT EDIT!
-# Generated by "Unix Makefiles" Generator, CMake Version 3.13
-
-CMakeFiles/cocktail.dir/cot_debug.c.o
- /tmp/tmp.RV7yFDR3Zp/Source/cot_debug.c
- /tmp/tmp.RV7yFDR3Zp/Source/cot_debug.h
-CMakeFiles/cocktail.dir/cot_lexer.c.o
- /tmp/tmp.RV7yFDR3Zp/Source/cocktail.h
- /tmp/tmp.RV7yFDR3Zp/Source/cot_debug.h
- /tmp/tmp.RV7yFDR3Zp/Source/cot_lexer.c
- /tmp/tmp.RV7yFDR3Zp/Source/cot_lexer.h
- /tmp/tmp.RV7yFDR3Zp/Source/cot_type.h
- /tmp/tmp.RV7yFDR3Zp/Source/cot_util.h
-CMakeFiles/cocktail.dir/cot_parser.c.o
- /tmp/tmp.RV7yFDR3Zp/Source/cocktail.h
- /tmp/tmp.RV7yFDR3Zp/Source/cot_parser.c
- /tmp/tmp.RV7yFDR3Zp/Source/cot_parser.h
- /tmp/tmp.RV7yFDR3Zp/Source/cot_type.h
- /tmp/tmp.RV7yFDR3Zp/Source/cot_util.h
-CMakeFiles/cocktail.dir/cot_stack.c.o
- /tmp/tmp.RV7yFDR3Zp/Source/cot_stack.c
- /tmp/tmp.RV7yFDR3Zp/Source/cot_stack.h
- /tmp/tmp.RV7yFDR3Zp/Source/cot_type.h
- /tmp/tmp.RV7yFDR3Zp/Source/cot_util.h
-CMakeFiles/cocktail.dir/cot_util.c.o
- /tmp/tmp.RV7yFDR3Zp/Source/cot_type.h
- /tmp/tmp.RV7yFDR3Zp/Source/cot_util.c
- /tmp/tmp.RV7yFDR3Zp/Source/cot_util.h
-CMakeFiles/cocktail.dir/main.c.o
- /tmp/tmp.RV7yFDR3Zp/Source/cocktail.h
- /tmp/tmp.RV7yFDR3Zp/Source/cot_debug.h
- /tmp/tmp.RV7yFDR3Zp/Source/cot_lexer.h
- /tmp/tmp.RV7yFDR3Zp/Source/cot_parser.h
- /tmp/tmp.RV7yFDR3Zp/Source/cot_type.h
- /tmp/tmp.RV7yFDR3Zp/Source/cot_util.h
- /tmp/tmp.RV7yFDR3Zp/Source/main.c
