--- Source/cmake-build-debug/CMakeFiles/cocktail.dir/depend.make
@@ -1,36 +1,0 @@
-# CMAKE generated file: DO NOT EDIT!
-# Generated by "Unix Makefiles" Generator, CMake Version 3.13
-
-CMakeFiles/cocktail.dir/cot_debug.c.o: ../cot_debug.c
-CMakeFiles/cocktail.dir/cot_debug.c.o: ../cot_debug.h
-
-CMakeFiles/cocktail.dir/cot_lexer.c.o: ../cocktail.h
-CMakeFiles/cocktail.dir/cot_lexer.c.o: ../cot_debug.h
-CMakeFiles/cocktail.dir/cot_lexer.c.o: ../cot_lexer.c
-CMakeFiles/cocktail.dir/cot_lexer.c.o: ../cot_lexer.h
-CMakeFiles/cocktail.dir/cot_lexer.c.o: ../cot_type.h
-CMakeFiles/cocktail.dir/cot_lexer.c.o: ../cot_util.h
-
-CMakeFiles/cocktail.dir/cot_parser.c.o: ../cocktail.h
-CMakeFiles/cocktail.dir/cot_parser.c.o: ../cot_parser.c
-CMakeFiles/cocktail.dir/cot_parser.c.o: ../cot_parser.h
-CMakeFiles/cocktail.dir/cot_parser.c.o: ../cot_type.h
-CMakeFiles/cocktail.dir/cot_parser.c.o: ../cot_util.h
-
-CMakeFiles/cocktail.dir/cot_stack.c.o: ../cot_stack.c
-CMakeFiles/cocktail.dir/cot_stack.c.o: ../cot_stack.h
-CMakeFiles/cocktail.dir/cot_stack.c.o: ../cot_type.h
-CMakeFiles/cocktail.dir/cot_stack.c.o: ../cot_util.h
-
-CMakeFiles/cocktail.dir/cot_util.c.o: ../cot_type.h
-CMakeFiles/cocktail.dir/cot_util.c.o: ../cot_util.c
-CMakeFiles/cocktail.dir/cot_util.c.o: ../cot_util.h
-
-CMakeFiles/cocktail.dir/main.c.o: ../cocktail.h
-CMakeFiles/cocktail.dir/main.c.o: ../cot_debug.h
-CMakeFiles/cocktail.dir/main.c.o: ../cot_lexer.h
-CMakeFiles/cocktail.dir/main.c.o: ../cot_parser.h
-CMakeFiles/cocktail.dir/main.c.o: ../cot_type.h
-CMakeFiles/cocktail.dir/main.c.o: ../cot_util.h
-CMakeFiles/cocktail.dir/main.c.o: ../main.c
-
