--- Source/cmake-build-debug/CMakeFiles/cocktail.dir/flags.make
@@ -1,10 +1,0 @@
-# CMAKE generated file: DO NOT EDIT!
-# Generated by "Unix Makefiles" Generator, CMake Version 3.13
-
-# compile C with /usr/bin/gcc
-C_FLAGS = -g  
-
-C_DEFINES = 
-
-C_INCLUDES = 
-
