--- Source/cmake-build-debug/CMakeFiles/cocktail.dir/progress.make
@@ -1,8 +1,0 @@
-CMAKE_PROGRESS_1 = 1
-CMAKE_PROGRESS_2 = 2
-CMAKE_PROGRESS_3 = 3
-CMAKE_PROGRESS_4 = 4
-CMAKE_PROGRESS_5 = 5
-CMAKE_PROGRESS_6 = 6
-CMAKE_PROGRESS_7 = 7
-
