--- Source/cmake-build-debug/cocktail.cbp
@@ -1,118 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<CodeBlocks_project_file>
-	<FileVersion major="1" minor="6"/>
-	<Project>
-		<Option title="cocktail"/>
-		<Option makefile_is_custom="1"/>
-		<Option compiler="gcc"/>
-		<Option virtualFolders="CMake Files\;"/>
-		<Build>
-			<Target title="all">
-				<Option working_dir="/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug"/>
-				<Option type="4"/>
-				<MakeCommands>
-					<Build command="/usr/bin/make -j16 -f &quot;/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug/Makefile&quot;  VERBOSE=1 all"/>
-					<CompileFile command="/usr/bin/make -j16 -f &quot;/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug/Makefile&quot;  VERBOSE=1 &quot;$file&quot;"/>
-					<Clean command="/usr/bin/make -j16 -f &quot;/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug/Makefile&quot;  VERBOSE=1 clean"/>
-					<DistClean command="/usr/bin/make -j16 -f &quot;/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug/Makefile&quot;  VERBOSE=1 clean"/>
-				</MakeCommands>
-			</Target>
-			<Target title="rebuild_cache">
-				<Option working_dir="/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug"/>
-				<Option type="4"/>
-				<MakeCommands>
-					<Build command="/usr/bin/make -j16 -f &quot;/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug/Makefile&quot;  VERBOSE=1 rebuild_cache"/>
-					<CompileFile command="/usr/bin/make -j16 -f &quot;/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug/Makefile&quot;  VERBOSE=1 &quot;$file&quot;"/>
-					<Clean command="/usr/bin/make -j16 -f &quot;/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug/Makefile&quot;  VERBOSE=1 clean"/>
-					<DistClean command="/usr/bin/make -j16 -f &quot;/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug/Makefile&quot;  VERBOSE=1 clean"/>
-				</MakeCommands>
-			</Target>
-			<Target title="cocktail">
-				<Option output="/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug/cocktail" prefix_auto="0" extension_auto="0"/>
-				<Option working_dir="/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug"/>
-				<Option object_output="./"/>
-				<Option type="1"/>
-				<Option compiler="gcc"/>
-				<Compiler>
-					<Add directory="/usr/lib/gcc/x86_64-linux-gnu/8/include"/>
-					<Add directory="/usr/local/include"/>
-					<Add directory="/usr/lib/gcc/x86_64-linux-gnu/8/include-fixed"/>
-					<Add directory="/usr/include/x86_64-linux-gnu"/>
-					<Add directory="/usr/include"/>
-				</Compiler>
-				<MakeCommands>
-					<Build command="/usr/bin/make -j16 -f &quot;/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug/Makefile&quot;  VERBOSE=1 cocktail"/>
-					<CompileFile command="/usr/bin/make -j16 -f &quot;/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug/Makefile&quot;  VERBOSE=1 &quot;$file&quot;"/>
-					<Clean command="/usr/bin/make -j16 -f &quot;/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug/Makefile&quot;  VERBOSE=1 clean"/>
-					<DistClean command="/usr/bin/make -j16 -f &quot;/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug/Makefile&quot;  VERBOSE=1 clean"/>
-				</MakeCommands>
-			</Target>
-			<Target title="cocktail/fast">
-				<Option output="/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug/cocktail" prefix_auto="0" extension_auto="0"/>
-				<Option working_dir="/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug"/>
-				<Option object_output="./"/>
-				<Option type="1"/>
-				<Option compiler="gcc"/>
-				<Compiler>
-					<Add directory="/usr/lib/gcc/x86_64-linux-gnu/8/include"/>
-					<Add directory="/usr/local/include"/>
-					<Add directory="/usr/lib/gcc/x86_64-linux-gnu/8/include-fixed"/>
-					<Add directory="/usr/include/x86_64-linux-gnu"/>
-					<Add directory="/usr/include"/>
-				</Compiler>
-				<MakeCommands>
-					<Build command="/usr/bin/make -j16 -f &quot;/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug/Makefile&quot;  VERBOSE=1 cocktail/fast"/>
-					<CompileFile command="/usr/bin/make -j16 -f &quot;/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug/Makefile&quot;  VERBOSE=1 &quot;$file&quot;"/>
-					<Clean command="/usr/bin/make -j16 -f &quot;/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug/Makefile&quot;  VERBOSE=1 clean"/>
-					<DistClean command="/usr/bin/make -j16 -f &quot;/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug/Makefile&quot;  VERBOSE=1 clean"/>
-				</MakeCommands>
-			</Target>
-			<Target title="edit_cache">
-				<Option working_dir="/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug"/>
-				<Option type="4"/>
-				<MakeCommands>
-					<Build command="/usr/bin/make -j16 -f &quot;/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug/Makefile&quot;  VERBOSE=1 edit_cache"/>
-					<CompileFile command="/usr/bin/make -j16 -f &quot;/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug/Makefile&quot;  VERBOSE=1 &quot;$file&quot;"/>
-					<Clean command="/usr/bin/make -j16 -f &quot;/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug/Makefile&quot;  VERBOSE=1 clean"/>
-					<DistClean command="/usr/bin/make -j16 -f &quot;/tmp/tmp.RV7yFDR3Zp/Source/cmake-build-debug/Makefile&quot;  VERBOSE=1 clean"/>
-				</MakeCommands>
-			</Target>
-		</Build>
-		<Unit filename="/tmp/tmp.RV7yFDR3Zp/Source/cot_debug.c">
-			<Option target="cocktail"/>
-		</Unit>
-		<Unit filename="/tmp/tmp.RV7yFDR3Zp/Source/cot_debug.h">
-			<Option target="cocktail"/>
-		</Unit>
-		<Unit filename="/tmp/tmp.RV7yFDR3Zp/Source/cot_lexer.c">
-			<Option target="cocktail"/>
-		</Unit>
-		<Unit filename="/tmp/tmp.RV7yFDR3Zp/Source/cot_lexer.h">
-			<Option target="cocktail"/>
-		</Unit>
-		<Unit filename="/tmp/tmp.RV7yFDR3Zp/Source/cot_parser.c">
-			<Option target="cocktail"/>
-		</Unit>
-		<Unit filename="/tmp/tmp.RV7yFDR3Zp/Source/cot_parser.h">
-			<Option target="cocktail"/>
-		</Unit>
-		<Unit filename="/tmp/tmp.RV7yFDR3Zp/Source/cot_stack.c">
-			<Option target="cocktail"/>
-		</Unit>
-		<Unit filename="/tmp/tmp.RV7yFDR3Zp/Source/cot_stack.h">
-			<Option target="cocktail"/>
-		</Unit>
-		<Unit filename="/tmp/tmp.RV7yFDR3Zp/Source/cot_util.c">
-			<Option target="cocktail"/>
-		</Unit>
-		<Unit filename="/tmp/tmp.RV7yFDR3Zp/Source/cot_util.h">
-			<Option target="cocktail"/>
-		</Unit>
-		<Unit filename="/tmp/tmp.RV7yFDR3Zp/Source/main.c">
-			<Option target="cocktail"/>
-		</Unit>
-		<Unit filename="/tmp/tmp.RV7yFDR3Zp/Source/CMakeLists.txt">
-			<Option virtualFolder="CMake Files\"/>
-		</Unit>
-	</Project>
-</CodeBlocks_project_file>
