--- Source/cocktail.h
@@ -1,20 +1,0 @@
-/*
- * Created by Lettle on 2021/9/7.
- * QQ: 1071445082
- * github: https://github.com/Python-Lettle
- * gitee : https://gitee.com/lettle/
- */
-#ifndef COCKTAIL_COCKTAIL_H
-#define COCKTAIL_COCKTAIL_H
-#include "cot_util.h"
-
-/**********************************************************************
- * 全局变量
- **********************************************************************/
-
-// token首
-cot_token *cot_token_head;
-// 全局块
-cot_global_block *cot_global;
-
-#endif //COCKTAIL_COCKTAIL_H
