--- Source/cot_stack.c
@@ -1,8 +1,0 @@
-/*
- * Created by Lettle on 2021/8/14.
- * QQ: 1071445082
- * github: https://github.com/Python-Lettle
- * gitee : https://gitee.com/lettle/
- */
-
-#include "cot_stack.h"
